Move docs away from gulp and into its own script file
Let caller handle error from buildDocs function and fix kramdown link extension
